% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mars_cols.R
\docType{data}
\name{Mars_colors}
\alias{Mars_colors}
\title{Corporate Colors}
\format{
An object of class \code{character} of length 20.
}
\usage{
Mars_colors
}
\description{
This is simply a list of the web color codes.
}
\keyword{datasets}
