% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mars_cols.R
\name{Mars_cols}
\alias{Mars_cols}
\title{Call specified color}
\usage{
Mars_cols(...)
}
\arguments{
\item{color}{Name of a color, or colors in the Mars_cols list}
}
\value{
Name and web code of the specified color
}
\description{
This function outputs the name and web code for a called color.
It can be used in ggplot graph code to use discrete colors. There are
a total of 20 colors in the list. Options are "blue", "teal", "yellow",
"skyblue", "green", "purple", "pink", "red", "orange", "grey", "blue50",
"teal50", "yellow50", "skyblue50","green50", "purple50", "pink50",
"red50", "orange50", and "white".
}
