% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mars_cols.R
\name{Mars_pal}
\alias{Mars_pal}
\title{Call specified color palette}
\usage{
Mars_pal(palette = "all", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a color palette}
}
\value{
Name and web code of the specified color palette
}
\description{
This function outputs the name and web code for a called color palette.
It reverses the order of the palatte as well, so gradients can be made of
each palette.
Palettes include "all", "cool", "warm", "solids", and "hues".
}
