% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mars_cols.R
\docType{data}
\name{Mars_palettes}
\alias{Mars_palettes}
\title{Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
Mars_palettes
}
\description{
This is a list of various palettes generated using the 20
colors.
Palettes include "all", "cool", "warm", "solids", and "hues".
}
\keyword{datasets}
