% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mars_cols.R
\name{scale_color_Mars}
\alias{scale_color_Mars}
\title{Scale color for ggplot inputs}
\usage{
scale_color_Mars(palette = "all", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a Mars color palette}
}
\value{
Scale color
}
\description{
This function creates a discrete and continuous color scale
of a specified palette to use in ggplot code.
Palettes include "all", "cool", "warm", "solids", and "hues".
}
