% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mars_cols.R
\name{scale_fill_Mars}
\alias{scale_fill_Mars}
\title{Scale fill for ggplot inputs}
\usage{
scale_fill_Mars(palette = "all", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of a Mars color palette}
}
\value{
Scale fill
}
\description{
This function creates a scale fill of a specified
palette to use in ggplot code.
Palettes include "all", "cool", "warm", "solids", and "hues".
}
